+++ src/main/java/project/cashier/ICashierManager.java
@@ -1,9 +1,10 @@
 package project.cashier;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ICashierManager {
     List<Cashier> getCashiers();
     void addCashier(Cashier cashier);
-    double getCashierSalary(long cashierId);
+    BigDecimal getCashierSalary(long cashierId);
 }
